Add files via upload
This is a code which is essentially a very rudimentary attempt at creating a spelling checker in python using the textblob library. This library uses NLTK package from python and allows the user to input an incorrectly spelt word and then it returns the correct spelling of the word. This code was inspired from a JavaTpoint tutorial.